Fixed Archive scrolling and other improvements
