+++ src/Outreach/Outreach.scss
@@ -5,6 +5,12 @@
     justify-content: center;
     flex-wrap: wrap-reverse;
     cursor: pointer;
+    img {
+        box-shadow: 0px 0px 10px $school-gold;
+    }
+    img:hover {
+        box-shadow: 0px 0px 15px $main-white;
+    }
 }
 
 .containers {
